src/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
fastapi==0.105.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
